requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.13.0
pandas>=2.0.0
scikit-learn>=1.3.0
tqdm>=4.66.0
jupyter>=1.0.0
ipywidgets>=8.1.0
datasets>=2.14.0
huggingface-hub>=0.20.0
python-dotenv>=1.0.0
optuna>=3.4.0
umap-learn>=0.5.4
gradio>=4.0.0
pytorch-fid>=0.3.0
lpips>=0.1.4
